requirements.txt
anyio==3.6.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dearpygui==1.6.2
deprecation==2.1.0
Flask==2.1.2
flask-oidc==1.4.0
h11==0.12.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
oauth2client==4.1.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
postgrest-py==0.10.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
